dom/workers/test/test_dynamicImport_and_terminate.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Worker create script loader failure</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

document.addEventListener("DOMContentLoaded", () => {
  const worker = new Worker("worker_dynamicImport.mjs", {"type": "module"});
  setTimeout(() => {
    worker.terminate();
    ok(true, "done");
    SimpleTest.finish();
  }, 0);
});

</script>
</pre>
</body>
</html>
